+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/DisplayProperties.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 /**
- * OID4VCI: OPTIONAL. An array of objects, where each object contains display properties of a Credential Issuer for a
- * certain language.
+ * OID4VCI: OPTIONAL. Array of objects, where each object contains the display properties of the supported Credential
+ * for a certain language.
  */
 @Serializable
 data class DisplayProperties(
@@ -21,4 +21,36 @@ data class DisplayProperties(
      */
     @SerialName("locale")
     val locale: String? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. Object with information about the logo of the Credential.
+     */
+    @SerialName("logo")
+    val logo: DisplayLogoProperties? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. String value of a description of the Credential.
+     */
+    @SerialName("description")
+    val description: String? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. String value of a background color of the Credential represented as numerical color values
+     * defined in CSS Color Module Level 37.
+     */
+    @SerialName("background_color")
+    val backgroundColor: String? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. Object with information about the background image of the Credential.
+     */
+    @SerialName("background_image")
+    val backgroundImage: DisplayLogoProperties? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. String value of a text color of the Credential represented as numerical color values defined
+     * in CSS Color Module Level 37
+     */
+    @SerialName("text_color")
+    val textColor: String? = null,
 )